campus_alert_agent/config.py
"""Configuration module for campus_alert_agent."""
import os
from dotenv import load_dotenv

load_dotenv()

GEMINI_API_KEY = os.environ.get("GEMINI_API_KEY", "")
ALERT_EMAIL_TO = os.environ.get("ALERT_EMAIL_TO", "")
ALERT_EMAIL_FROM = os.environ.get("ALERT_EMAIL_FROM", "")
GMAIL_APP_PASSWORD = os.environ.get("GMAIL_APP_PASSWORD", "")
TELEGRAM_BOT_TOKEN = os.environ.get("TELEGRAM_BOT_TOKEN", "")
TELEGRAM_CHAT_ID = os.environ.get("TELEGRAM_CHAT_ID", "")

CHECK_INTERVAL_MINUTES = 60

SOURCES = [
    "https://www.anthropic.com/news",
    "https://www.anthropic.com/careers",
]

GEMINI_MODEL = "gemini-1.5-flash"
MAX_TEXT_CHARS = 6000
